parquet: clean up API (ColumnChunkReader)
Main points:

- Concrete ColumnChunkReaders are no longer exported
- Replace concrete constructor functions for ColumnChunkReader with a single
  NewColumnChunkReader
- Do not use pointers where not necessary
- Remove NextPage() from ColumnChunkReader (will be added later if needed)
- update parqueteur to use new API

Commit to remove duplicated code will follow.
